+++ helper/acvpkeys/generate.sh
@@ -5,4 +5,4 @@
 #
 
 openssl genrsa -aes256 -out privkey-rsa.pem 4096
-openssl req -config openssl-req.cnf -new -key privkey-rsa.pem -out csr-rsa.pem -sha256
+openssl req -config openssl-req.cnf -new -key privkey-rsa.pem -out atsec_corp-FIRSTNAME_LASTNAME-Demo.pem -sha256